pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "movie-revenue-prediction"
version = "0.1.0"
description = "Movie revenue prediction using TMDB data with feature engineering, traditional ML models, neural networks and an ensemble."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Nicolae" }
]
license = { text = "MIT" }

dependencies = [
  "numpy>=1.23",
  "pandas>=1.5",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "tqdm>=4.65",
  "joblib>=1.3",
  "scikit-learn>=1.3",
  "xgboost>=1.7",
  "scikeras>=0.10",
  "tensorflow>=2.13",
  "pyarrow>=12.0",
  "jupyter>=1.0",
  "notebook>=7.0",
  "ipykernel>=6.25",
  "matplotlib-inline>=0.1",
]

[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "pytest",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
